Update imageCapture.cpp
Added the confidence level calculation. Also, fixed a bug in which the blink duration was not updated in real time. It was only updated after the blink is complete. I made it so it updates as soon as the eyes are closed. Also added the head orientation, basic model. This needs to be improved right now it is not very reliable.